Move macro unit tests outside of components
