+++ test/unit/component-helper.js
@@ -11,7 +11,6 @@ const nunjucks = nunjucksConfig(null, {
 })
 
 const COMPONENTS_PATH = '_components/'
-const MACROS_PATH = '_macros/'
 const EXT = 'njk'
 
 const normaliseHtml = (string) => {
@@ -23,52 +22,6 @@ const normaliseHtml = (string) => {
   return html.prettyPrint(beautiful, {
     indent_size: 2,
   })
-}
-
-function getMacros (fileName) {
-  const filePath = `${MACROS_PATH}${fileName}.${EXT}`
-
-  function renderMacro (name, caller, props, ...rest) {
-    const importString = `{% from "${filePath}" import ${name} %}`
-    const args = rest.length ? `,${[...rest].map(JSON.stringify)}` : ''
-    let macroOutput
-
-    if (caller) {
-      macroOutput = nunjucks.renderString(`
-        ${importString}
-        {% call ${name}(${JSON.stringify(props, undefined, 1)}${args}) %}
-          ${caller.join(' ')}
-        {% endcall %}
-      `)
-    } else {
-      macroOutput = nunjucks.renderString(`${importString} {{ ${name}(${JSON.stringify(props, undefined, 1)}${args}) }}`)
-    }
-
-    return normaliseHtml(macroOutput)
-  }
-
-  return {
-    render (macroName, props, ...rest) {
-      return renderMacro(macroName, null, props, ...rest)
-    },
-
-    renderWithCaller (macroName, caller, props, ...rest) {
-      return renderMacro(macroName, caller, props, ...rest)
-    },
-
-    renderToDom (macroName, props, ...rest) {
-      const macroOutput = this.render(macroName, props, ...rest)
-      return (new JSDOM(macroOutput)).window.document.body.firstElementChild
-    },
-
-    renderWithCallerToDom (macroName, props, ...rest) {
-      return function (...caller) {
-        if (!caller.length) { return null }
-        const macroOutput = this.renderWithCaller(macroName, caller, props, ...rest)
-        return (new JSDOM(macroOutput)).window.document.body.firstElementChild
-      }.bind(this)
-    },
-  }
 }
 
 const renderComponent = (name, input) => {
@@ -100,8 +53,8 @@ function domTokenToArray (obj) {
 }
 
 module.exports = {
+  normaliseHtml,
   expectComponent,
-  getMacros,
   renderComponentToDom,
   domTokenToArray,
 }
