+++ test/unit/macros/local-header.js
@@ -1,4 +1,4 @@
-const { getMacros } = require('~/test/unit/component-helper')
+const { getMacros } = require('~/test/unit/macro-helper')
 const commonMacros = getMacros('common')
 
 describe('LocalHeader macro', () => {
